refactor: final path refactor commit
