+++ src/components/Order/components/send_order.tsx
@@ -1,10 +1,7 @@
 import Button from '@mui/material/Button';
 import { toast } from 'react-toastify';
 
-import {
-  useDeleteCartMutation,
-  useGetActiveCartQuery,
-} from '../../../services/api';
+import { useDeleteCartMutation, useGetActiveCartQuery } from '@/services/api';
 
 const SendOrder = ({ onCartDeleted }: { onCartDeleted: () => void }) => {
   const { data: cart } = useGetActiveCartQuery({});
