+++ src/pages/profile/auth-layout/address-form/baseSchema.ts
@@ -3,7 +3,7 @@ import * as yup from 'yup';
 import {
   onlyLettersRegex,
   postalCodeRegex,
-} from '../../../login-register/register/interfaces';
+} from '@/pages/login-register/register/interfaces';
 
 const baseSchema = yup.object().shape({
   country: yup.string().required('Country is required'),
