+++ src/pages/profile/auth-layout/address-form/existed-address/ExistedAddress.tsx
@@ -2,13 +2,14 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { FormProvider, useForm } from 'react-hook-form';
 import { toast } from 'react-toastify';
 
-import { useAppDispatch } from '../../../../../redux/hooks';
-import { setCustomer } from '../../../../../redux/slices/authSlice';
-import { useUpdateProfileMutation } from '../../../../../services/api';
+import { useAppDispatch } from '@/redux/hooks';
+import { setCustomer } from '@/redux/slices/authSlice';
+import { useUpdateProfileMutation } from '@/services/api';
 import {
   getCountryCodeByName,
   getCountryNameByCode,
-} from '../../../../../utils/getCountryNameByCode';
+} from '@/utils/getCountryNameByCode';
+
 import AddressForm, { AddressFormValues } from '../AddressForm';
 import baseSchema from '../baseSchema';
