+++ src/pages/profile/auth-layout/address-list/AddressesList.tsx
@@ -13,9 +13,9 @@ import {
 } from '@mui/material';
 
 import Title from '@/components/title/Title';
+import { Address } from '@/interfaces/customerDeteils.interface';
 import { getCountryNameByCode } from '@/utils/getCountryNameByCode';
 
-import { Address } from '../../../../interfaces/customerDeteils.interface';
 import ExistedAddress from '../address-form/existed-address/ExistedAddress';
 
 interface IProps {
